Add Code of Conduct (#32)
* Move 'Contribute' instructions into separate doc

We are about to add a `CONTRIBUTING.md` doc that will have the following piece, taken from the `README.md`

```
## Contribute

The main purpose of this repository is to share Facebook's implementation of an emitter. Please see React's [contributing article](https://github.com/facebook/react/blob/master/CONTRIBUTING.md), which generally applies to `fbemitter`, if you are interested in submitting a pull request.
```

* Add `CONTRIBUTING.md`

Even though it just points to the React 'Contributing' article, this is where people will conventionally expect to find details about Contributing.

* Link to COC in `CONTRIBUTING.md`

We are about to add a Code of Conduct, and it makes sense to surface that in the `CONTRIBUTING.md`.

* Add `CODE_OF_CONDUCT.md`

In the past Facebook didn't promote including a Code of Conduct when creating new projects, and many projects skipped this important document. Let's fix it. :)

Even though this project is not super active, it still makes sense to have a Code of Conduct.

**why make this change?:**
Facebook Open Source provides a Code of Conduct statement for all
projects to follow, to promote a welcoming and safe open source community.

Exposing the COC via a separate markdown file is a standard being
promoted by Github via the Community Profile in order to meet their Open
Source Guide's recommended community standards.

As you can see, adding this file will improve [the emitter community profile](https://github.com/facebook/emitter/community)
checklist and increase the visibility of our COC.

**test plan:**
Viewing it on my branch -
(Flarnie will insert screenshots)

**issue:**
internal task t23481323